Move port bounds to the endpoint worker
The upper and lower bounds are only used by the endpoint worker, so it
is now removed from other types. The new_worker function now takes
mutable references to these bounds and the generation of independent
port ranges is delegated to this function. This also means the workers
must now be created by the parent process before forking itself in order
to have one source of truth for the ranges (but this should be fine,
since in the parent the worker goes out of scope right after forking).