BPF worker
BPF program worker to simulate contention on a tracepoint. The
implementation benefits from aya types, but uses raw syscalls to do
actual job, since we don't want to depend on any external BPF programs
and generate everything on the fly. The resulting programs are simplest
possible, which is enough to simulate a contention.